// apps/api-core/src/resources/cacheKeyName.ts
export enum CacheKeyName {
  CRON_SETTINGS_LIST = "CRON_SETTINGS_LIST",
  ENV_DELAY = "ENV_DELAY",
  GLOBAL_INFO = "GLOBAL_INFO",
  PROXY_CONFIG_BY_PROVIDER_ID = "PROXY_CONFIG_BY_PROVIDER_ID",
  FILTER_CRON_DETAILS = "FILTER_CRON_DETAILS",
  SLOW_CRON_DETAILS = "SLOW_CRON_DETAILS",
  SCRAPE_CRON_DETAILS = "SCRAPE_CRON_DETAILS",
  PROXY_CONFIG_BY_PRIORITY = "PROXY_CONFIG_BY_PRIORITY",
  GET_ALL_PROXIES = "GET_ALL_PROXIES",
  PROXY_SWITCH_CRON_DETAILS = "PROXY_SWITCH_CRON_DETAILS",
  PROXY_SWITCH_CURRENT_CRON_ID = "",
  PROXY_SWITCH_CURRENT_PROXY_PRIORITY = "",
  _422_RUNNING_CACHE = "422_RUNNING_CACHE",
  SMART_PROXY_WEB_COUNTER_CACHE = "SMART_PROXY_WEB_COUNTER_CACHE",
  SCRAPING_BEE_NON_JS_COUNTER_CACHE = "SCRAPING_BEE_NON_JS_COUNTER_CACHE",
  SCRAPING_BEE_JS_COUNTER_CACHE = "SCRAPING_BEE_JS_COUNTER_CACHE",
  BRIGHTDATA_COUNTER_CACHE = "BRIGHTDATA_COUNTER_CACHE",
}